requirements.txt
pytrends
pandas
gspread
oauth2client
google-auth